refactor(frontend-component-tierList)
