+++ app/components/tierList.jsx
@@ -1,4 +1,4 @@
-import { useState, useEffect } from 'react';
+import { useState } from 'react';
 import {
   DndContext,
   pointerWithin,
@@ -9,13 +9,19 @@ import {
   MeasuringStrategy,
 } from '@dnd-kit/core';
 import { restrictToWindowEdges } from '@dnd-kit/modifiers';
-import { arrayMove, verticalListSortingStrategy } from '@dnd-kit/sortable';
+import { verticalListSortingStrategy } from '@dnd-kit/sortable';
 import { SortableContext } from '@dnd-kit/sortable';
 import TierListRow from '@/app/components/tierListRow';
 import DragOverLayTierList from '@/app/components/dragOverLayTierList';
 import Spawner from '@/app/components/spawner';
 import { Button } from './button';
-import Swal from 'sweetalert2';
+import {
+  handleDragOver,
+  handleDragEnd,
+  handleDragStart,
+  handleAddRow,
+  handleRemoveElement,
+} from '@/utils/tierList/handler';
 
 const measuringConfig = {
   droppable: {
@@ -31,372 +37,16 @@ export default function TierList({
   tierListId,
 }) {
   const [activeId, setActiveId] = useState(null); // for over lay
-  const defaultRow = [
-    {
-      label: 'All of my heart',
-      bgColor: '#FAFEFF',
-    },
-    {
-      label: 'Part of my heart',
-      bgColor: '#F1EEE7',
-    },
-    {
-      label: 'In my heart',
-      bgColor: '#FAE6BE',
-    },
-    {
-      label: 'Love that',
-      bgColor: '#F2FABE',
-    },
-    {
-      label: 'Ahh Ha',
-      bgColor: '#D8FABE',
-    },
-    {
-      label: 'Ummm',
-      bgColor: '#BEFAE4',
-    },
-    {
-      label: "What that's",
-      bgColor: '#BEECFA',
-    },
-    {
-      label: 'Nope',
-      bgColor: '#BED2FA',
-    },
-  ];
+  const itemsLength = items.length;
+  const lastIndexOriginal = itemsLength - 2; // index of last row
 
-  const removeElement = (id, parentIndex) => {
-    setItems((prev) => {
-      return [
-        ...prev.slice(0, parentIndex),
-        {
-          ...prev[parentIndex],
-          elements: prev[parentIndex].elements.filter(
-            (element) => element.id !== id
-          ),
-        },
-        ...prev.slice(parentIndex + 1, prev.length),
-      ];
-    });
-  };
-
-  const handleRemoveElement = (id, parentIndex) => {
-    Swal.fire({
-      title: 'Are you sure?',
-      color: '#A73440',
-      text: 'Do you really want to delete this Image?',
-      iconHtml:
-        '<Image src="/iconTrash.svg" width=65px height=65px alt="delete icon" />',
-      showCancelButton: true,
-      confirmButtonColor: '#a73440',
-      cancelButtonColor: '#DE717C',
-      confirmButtonText: 'Delete',
-      buttonsStyling: false,
-      customClass: {
-        popup:
-          'flex flex-col gap-[15px] bg-peach border border-cream rounded-lg',
-        title: 'p-0',
-        htmlContainer: 'm-0',
-        icon: 'border-0',
-        actions: 'flex flex-col gap-[15px] w-1/2',
-        confirmButton:
-          'bg-winered py-2 text-white font-bold w-full rounded-lg shadow-lg border border-[#FAFEFF]',
-        cancelButton:
-          'bg-cherry py-2 text-white font-bold w-full rounded-lg shadow-lg border border-[#FAFEFF]',
-      },
-    }).then((result) => {
-      if (result.isConfirmed) {
-        removeElement(id, parentIndex);
-        Swal.mixin({
-          toast: true,
-          position: 'top-end',
-          showConfirmButton: false,
-          timer: 2000,
-          timerProgressBar: true,
-          didOpen: (toast) => {
-            toast.onmouseenter = Swal.stopTimer;
-            toast.onmouseleave = Swal.resumeTimer;
-          },
-        }).fire({
-          icon: 'success',
-          title: 'Delete successfully',
-        });
-      }
-    });
-  };
-
-  function findContainer(id) {
-    if (items.find((container) => container.id == id)) {
-      return id;
-    }
-
-    return items.find((row) =>
-      row.elements.find((element) => element.id == id)
-    );
-  }
-
-  function handleDragStart(event) {
-    setActiveId(event.active.id);
-  }
-
-  function handleDragOver(event) {
-    const { active, over } = event;
-    const { id } = active;
-    const overId = over?.id;
-
-    // Find the containers
-    const activeContainer = findContainer(id);
-    const overContainer = findContainer(overId);
-    const isSpawner = event.over?.id == 'spawner';
-
-    if (
-      !activeContainer ||
-      !overContainer ||
-      activeContainer === overContainer
-    ) {
-      return;
-    }
-
-    // drag element to empty row
-    if (
-      typeof overContainer === 'string' &&
-      typeof activeContainer === 'object'
-    ) {
-      const activeContainerIndex = items.findIndex(
-        (row) => row.id == activeContainer.id
-      );
-      const activeElementIndex = event.active.data.current.sortable.index;
-
-      // drag to spawner row
-      if (isSpawner && items[items.length - 1].elements.length == 0) {
-        setItems((prev) => {
-          return [
-            ...prev.slice(0, activeContainerIndex),
-            {
-              ...prev[activeContainerIndex],
-              elements: prev[activeContainerIndex].elements.filter(
-                (element) => element.id !== id
-              ),
-            },
-            ...prev.slice(activeContainerIndex + 1, -1),
-            {
-              ...prev[prev.length - 1],
-              elements: [
-                prev[activeContainerIndex].elements[activeElementIndex],
-              ],
-            },
-          ];
-        });
-        return;
-      }
-      if (!isSpawner) {
-        const overContainerIndex = event.over.data.current.sortable.index;
-        if (items[overContainerIndex].elements.length == 0) {
-          setItems((prev) => {
-            const add = {
-              ...prev[overContainerIndex],
-              ['elements']: [
-                prev[activeContainerIndex].elements[activeElementIndex],
-              ],
-            };
-
-            const remove = {
-              ...prev[activeContainerIndex],
-              ['elements']: prev[activeContainerIndex].elements.filter(
-                (element) => element.id !== id
-              ),
-            };
-
-            if (activeContainerIndex > overContainerIndex) {
-              return [
-                ...prev.slice(0, overContainerIndex),
-                add,
-                ...prev.slice(overContainerIndex + 1, activeContainerIndex),
-                remove,
-                ...prev.slice(activeContainerIndex + 1, prev.length),
-              ];
-            } else {
-              return [
-                ...prev.slice(0, activeContainerIndex),
-                remove,
-                ...prev.slice(activeContainerIndex + 1, overContainerIndex),
-                add,
-                ...prev.slice(overContainerIndex + 1, prev.length),
-              ];
-            }
-          });
-        }
-      }
-
-      return;
-    }
-
-    // swap row
-    if (
-      typeof activeContainer === 'string' &&
-      (overContainer == 'spawner' || overContainer?.id == 'spawner')
-    ) {
-      return;
-    }
-    if (
-      typeof activeContainer === 'string' &&
-      !(overContainer == 'spawner' || overContainer?.id == 'spawner')
-    ) {
-      const activeRowIndex = event.active.data.current.sortable.index;
-      const overRowIndex =
-        typeof overContainer === 'string'
-          ? event.over.data.current.sortable.index
-          : activeRowIndex;
-      setItems((prev) => {
-        return arrayMove(prev, activeRowIndex, overRowIndex);
-      });
-      return;
-    }
-
-    // move element to difference row
-    const activeContainerIndex = items?.findIndex(
-      (container) => container.id == activeContainer.id
-    );
-    let overContainerIndex = items?.findIndex(
-      (container) =>
-        container.elements?.filter((element) => element.id == overId).length ==
-        1
-    );
-    const activeIndex = event.active.data.current?.sortable?.index;
-    const overIndex = event.over.data.current?.sortable?.index;
-    if (
-      !items[overContainerIndex] ||
-      !items[overContainerIndex].elements ||
-      !items[overContainerIndex].label ||
-      (activeIndex != 0 && !activeIndex) ||
-      (overIndex != 0 && !overIndex) ||
-      !overId
-    ) {
-      return;
-    }
-    setItems((prev) => {
-      const add = prev[activeContainerIndex].elements[activeIndex];
-      if (overContainerIndex < activeContainerIndex) {
-        return [
-          ...prev.slice(0, overContainerIndex),
-          {
-            ...prev[overContainerIndex],
-            elements: [
-              ...prev[overContainerIndex].elements.slice(0, overIndex),
-              add,
-              ...prev[overContainerIndex].elements.slice(
-                overIndex,
-                prev[overContainerIndex].elements.length
-              ),
-            ],
-          },
-          ...prev.slice(overContainerIndex + 1, activeContainerIndex),
-          {
-            ...prev[activeContainerIndex],
-            elements: [
-              ...prev[activeContainerIndex].elements.filter(
-                (element) => element.id !== active.id
-              ),
-            ],
-          },
-          ...prev.slice(activeContainerIndex + 1, prev.length),
-        ];
-      } else {
-        return [
-          ...prev.slice(0, activeContainerIndex),
-          {
-            ...prev[activeContainerIndex],
-            elements: [
-              ...prev[activeContainerIndex].elements.filter(
-                (element) => element.id !== active.id
-              ),
-            ],
-          },
-          ...prev.slice(activeContainerIndex + 1, overContainerIndex),
-          {
-            ...prev[overContainerIndex],
-            elements: [
-              ...prev[overContainerIndex].elements.slice(0, overIndex),
-              add,
-              ...prev[overContainerIndex].elements.slice(
-                overIndex,
-                prev[overContainerIndex].elements.length
-              ),
-            ],
-          },
-          ...prev.slice(overContainerIndex + 1, prev.length),
-        ];
-      }
-    });
-  }
-
-  function handleDragEnd(event) {
-    const { active, over } = event;
-    const { id } = active;
-    const overId = over?.id;
-
-    const activeContainer = findContainer(id);
-    const overContainer = findContainer(overId);
-
-    if (
-      !activeContainer ||
-      !overContainer ||
-      activeContainer !== overContainer ||
-      typeof activeContainer === 'string'
-    ) {
-      return;
-    }
-
-    const activeIndex = activeContainer?.elements?.findIndex(
-      (element) => element.id == active.id
-    );
-
-    const overIndex = overContainer?.elements?.findIndex(
-      (element) => element.id == overId
-    );
-
-    const containerIndex = items?.findIndex(
-      (container) =>
-        container.elements.filter((element) => element.id == overId).length == 1
-    );
-
-    setItems((prev) => [
-      ...prev.slice(0, containerIndex),
-      {
-        ...prev[containerIndex],
-        elements: arrayMove(
-          prev[containerIndex].elements,
-          activeIndex,
-          overIndex
-        ),
-      },
-      ...prev.slice(containerIndex + 1, prev.length),
-    ]);
-    setActiveId(null);
-  }
-
-  const lastIndexOriginal = 1; // count when query
-  const handleAddRow = () => {
-    setItems([
-      ...items.slice(0, -1),
-      {
-        id: `container${items.length}`,
-        label: defaultRow[items.length - 1].label,
-        color: defaultRow[items.length - 1].bgColor,
-        elements: [],
-      },
-      items[items.length - 1],
-    ]);
-  };
-
+  // sensor
   const mouseSensor = useSensor(MouseSensor, {
     activationConstraint: {
       delay: 200, // millisec
       tolerance: 2,
     },
   }); // Initialize mouse sensor
-
   const touchSensor = useSensor(TouchSensor, {
     activationConstraint: {
       delay: 200, // millisec
@@ -404,15 +54,20 @@ export default function TierList({
     },
   }); // Initialize touch sensor
   const sensors = useSensors(touchSensor, mouseSensor);
-  const itemsLength = items.length;
 
   return (
     <DndContext
       id={tierListId}
       collisionDetection={pointerWithin}
-      onDragStart={handleDragStart}
-      onDragOver={handleDragOver}
-      onDragEnd={handleDragEnd}
+      onDragStart={(e) => {
+        handleDragStart(e, setActiveId);
+      }}
+      onDragOver={(e) => {
+        handleDragOver(e, setItems, items);
+      }}
+      onDragEnd={(e) => {
+        handleDragEnd(e, setItems, items, setActiveId);
+      }}
       measuring={measuringConfig}
       modifiers={[restrictToWindowEdges]}
       autoScroll={true}
@@ -423,21 +78,19 @@ export default function TierList({
           {/* each row of tier list part */}
           {items.map((row, index) => {
             // spawner part
-            if (index == items.length - 1) {
+            if (index == itemsLength - 1) {
               return (
-                // isEditable && (
                 <Spawner
                   setItems={setItems}
-                  rows={items}
                   isEditable={isEditable}
                   key={row.id}
                   items={items}
-                  handleRemoveElement={handleRemoveElement}
+                  handleRemoveElement={(id, parentIndex) => {
+                    handleRemoveElement(id, parentIndex, setItems);
+                  }}
                 />
-                // )
               );
             }
-
             // each row in tier list
             return (
               <TierListRow
@@ -448,7 +101,9 @@ export default function TierList({
                 key={row.id}
                 row={row}
                 rowIndex={index}
-                handleRemoveElement={handleRemoveElement}
+                handleRemoveElement={(id, parentIndex) => {
+                  handleRemoveElement(id, parentIndex, setItems);
+                }}
               />
             );
           })}
@@ -459,7 +114,9 @@ export default function TierList({
           >
             <div className='w-[200px]'>
               <Button
-                onClick={handleAddRow}
+                onClick={() => {
+                  handleAddRow(items, setItems);
+                }}
                 type={'addtextbtn'}
                 text={'Add new Level'}
               />
