+++ app/components/tierListRow.jsx
@@ -14,8 +14,9 @@ export default function TierListRow(props) {
       row={row}
       isRow={true}
       className={`transition ${
-        rowIndex == props?.itemsLength - 2 &&
-        props?.lastIndexOriginal != rowIndex &&
+        rowIndex == props.lastIndexOriginal &&
+        props.lastIndexOriginal + 1 != rowIndex &&
+        !props?.isOverLay &&
         'animate-fadeIn'
       } flex h-[90px] w-full shrink-0 gap-[5px] rounded-xl border border-[#F1EEE7] bg-[#F1EEE7] pr-1 shadow-lg `}
     >
